Add new configurable options for Starting Chest, CobbleGen, and Protected Worlds:
This makes it possible for users to modify the starting contents of the island chest and the drops from the cobble generator without modifying the plugin.  The proteted-worlds option is for worlds folders that should not be removed with the sbcleanup command.

+++ src/SkyBlock/command/SkyBlockCleanup.php
@@ -5,6 +5,7 @@ namespace SkyBlock\command;
 
 use pocketmine\command\Command;
 use pocketmine\command\CommandSender;
+use pocketmine\command\ConsoleCommandSender;
 use pocketmine\level\Level;
 use pocketmine\utils\Config;
 use SkyBlock\SkyBlock;
@@ -24,43 +25,53 @@ class SkyBlockCleanup extends Command{
 	}
 
 	public function execute(CommandSender $sender, string $commandLabel, array $args){
-		$validWorlds = ["ender", "nether", "pvp", "pvpisland", "skyblock", "world"];
-		$userConfigs = new \RecursiveIteratorIterator(new \RecursiveDirectoryIterator($this->usersDir, \RecursiveDirectoryIterator::SKIP_DOTS),
-			\RecursiveIteratorIterator::CHILD_FIRST);
-		foreach($userConfigs as $configFile) {
-			if(!$configFile->isDir()){
-				$config = new Config($configFile->getRealPath(), Config::JSON);
-				$island = $config->get("island");
-				array_push($validWorlds, $island);
+		if($sender instanceof ConsoleCommandSender){
+			$validWorlds = $this->plugin->getConfig()->get("protected-worlds");
+			if(!empty($validWorlds)){
+				$counter = 0;
+				$userConfigs = new \RecursiveIteratorIterator(new \RecursiveDirectoryIterator($this->usersDir, \RecursiveDirectoryIterator::SKIP_DOTS),
+					\RecursiveIteratorIterator::CHILD_FIRST);
+				foreach($userConfigs as $configFile){
+					if(!$configFile->isDir()){
+						$config = new Config($configFile->getRealPath(), Config::JSON);
+						$island = $config->get("island");
+						array_push($validWorlds, $island);
+					}
+				}
+				$worldFolders = glob($this->worldsDir . "*");
+				foreach($worldFolders as $worldFolder){
+					$string = $this->worldsDir;
+					$worldName = str_replace($string, "", $worldFolder);
+					if(!in_array($worldName, $validWorlds)){
+						$world = $this->plugin->getServer()->getLevelByName($worldName);
+						if($world instanceof Level){
+							$this->plugin->getServer()->unloadLevel($world);
+						}
+						$config = $this->islandsDir . $worldName . ".json";
+
+						$files = new \RecursiveIteratorIterator(
+							new \RecursiveDirectoryIterator($worldFolder . DIRECTORY_SEPARATOR, \RecursiveDirectoryIterator::SKIP_DOTS),
+							\RecursiveIteratorIterator::CHILD_FIRST
+						);
+
+						foreach($files as $fileinfo){
+							$todo = ($fileinfo->isDir() ? 'rmdir' : 'unlink');
+							$todo($fileinfo->getRealPath());
+							$counter++;
+						}
+
+						rmdir($worldFolder . DIRECTORY_SEPARATOR);
+						if(is_file($config)){
+							unlink($config);
+						}
+					}
+				}
+				$sender->sendMessage("sbcleanup: Completed!  Removed $counter files and directories.");
+			}else{
+				$sender->sendMessage("The configuration file needs to be updated before using the sbcleanup command.");
 			}
+		} else {
+			$sender->sendMessage("This command must be run from the console.");
 		}
-		$worldFolders = glob($this->worldsDir . "*");
-		foreach($worldFolders as $worldFolder){
-			$string = $this->worldsDir;
-			$worldName = str_replace($string, "", $worldFolder);
-			if(!in_array($worldName, $validWorlds)){
-				$world = $this->plugin->getServer()->getLevelByName($worldName);
-				if($world instanceof Level) {
-					$this->plugin->getServer()->unloadLevel($world);
-				}
-				$config = $this->islandsDir . $worldName . ".json";
-
-				$files = new \RecursiveIteratorIterator(
-					new \RecursiveDirectoryIterator($worldFolder . DIRECTORY_SEPARATOR, \RecursiveDirectoryIterator::SKIP_DOTS),
-					\RecursiveIteratorIterator::CHILD_FIRST
-				);
-
-				foreach ($files as $fileinfo) {
-					$todo = ($fileinfo->isDir() ? 'rmdir' : 'unlink');
-					$todo($fileinfo->getRealPath());
-				}
-
-				rmdir($worldFolder . DIRECTORY_SEPARATOR);
-				if(is_file($config)){
-					unlink($config);
-				}
-			}
-		}
-
 	}
 }
